add post new animal
