+++ src/components/dashboard/PropertyData.vue
@@ -1,10 +1,12 @@
 <script setup>
 import CardDashboard from './CardDashboard.vue';
+import PropertyDataModal from './PropertyDataModal.vue';
 </script>
 
 <template>
   <div class="col-12 col-md-6 d-flex align-items-center">
     <CardDashboard />
+    <PropertyDataModal />
   </div>
 </template>
